% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{Arith,Spectra,Spectra-method}
\alias{Arith,Spectra,Spectra-method}
\title{Apply arithmetic operations on/between Spectra objects}
\usage{
Arith(e1, e2)
e1 * e2
e1 + 0.5
}
\arguments{
\item{e1}{spectra object}

\item{e2}{spectra object}
}
\description{
Methods definig Arithmetic operations between two Spectra objects e1 and e2 or one
Spectra object e1 and a numeric value.
}
\details{
These methods allow performing arithmetic operations involving Spectra objects.
}
\seealso{
\code{\link{Arith}}
}

