% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{Arith,Spectra,Spectra-method}
\alias{Arith,Spectra,Spectra-method}
\title{Group Generic Functions}
\usage{
Arith(x, ...)
}
\arguments{
\item{...}{spectra or numeric objec}

\item{x}{spectra object}
}
\description{
Function takes out of a spectra objects from one to another one or from spectra object to numeric objects
}
\details{
Constructs a rectangle of sp::Lines using the bounding box of a Spatial object
}

