% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{$,Spectra-method}
\alias{$,Spectra-method}
\title{Extract or replace parts of a spectra object}
\usage{
x[i] <- spc.example_spectra()
x[i, j, ...] <- spc.example_spectra()
x[[i]] <- spc.example_spectra()
x$i <- spc.example_spectra()
}
\arguments{
\item{Spectra}{object from which to extract element(s) or in which to replace element(s)}
}
\description{
Operators acting on  spectral objects  to extract or replace parts
}
\examples{
 sp<-spc.example_spectra()
 # spc.colnames() is used to show that anap_300 or anap_345 is colon ,  
 spc.colnames(sp)
 sp$anap_300
 sp["anap_345"]
 


}

