% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{dim,Spectra-method}
\alias{dim,Spectra-method}
\title{Dimensions of a \code{Spectra} object.}
\usage{
\S4method{dim}{Spectra}(x)
}
\arguments{
\item{x}{A \code{Spectra} object}
}
\value{
Returns a numeric vector containing \code{nrow} and \code{ncol} of the \code{Spectra} object.
}
\description{
Gives number of dimension of a \code{Spectra} object
}
\examples{
sp<-spc.example_spectra()
dim(sp)

}
