% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\name{h2d}
\alias{h2d}
\title{Get header for data}
\usage{
spc.data2header(object,headerfield,dataname,compress=TRUE,...)
}
\arguments{
\item{object}{A \code{spclist} object}

\item{headerfield}{data column}

\item{dataname}{list \code{spclist} object}
}
\value{
object of class \code{spclist}
}
\description{
Get  the header for data of each element  with a column
}
\details{
If header element has length >1, its type is checked. If it is "character",
its elements will be pasted using paste(...,collapse="|"). If it is another 
type, only the first element will be taken.
}

