% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{head,Spectra-method}
\alias{head,Spectra-method}
\title{Return the first or last part of a Spectra object}
\usage{
head(x)
}
\arguments{
\item{x}{a specrta object}
}
\value{
Returns a matrix (Spectral data)
}
\description{
Return the first or last parts of a spectral object
}
\examples{
x <- spc.example_spectra()
head(x)

 
}

