% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{mat_identify}
\alias{mat_identify}
\title{Gives logical vector after clicking \code{Spectra} object in plot}
\usage{
mat_identify(x, y, ...)
}
\description{
when click on graph to select a \code{Spectra} adn print which row select,
 click Esc to quit after that, it returns logical vector
 @usage 
 spc.select(x, y, ...)
 @param x and y a \code{Spectra} object
 @param ... arguments to be passed to or from other methods.
}
\examples{
sp=spc.example_spectra()
spc.plot(sp)
spc.select(sp)
}
