% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\docType{methods}
\name{names,SpcList-method}
\alias{names,SpcList-method}
\title{The Names of a \code{Spclist} object}
\usage{
names(x)
}
\arguments{
\item{x}{a \code{Spclist} object}
}
\description{
Retrieve  the names of \code{Spclist} object
}
\examples{

sp <- spc.example_spectra() 
BL = spc.makeSpcList(sp,"STATION")
names(BL)

}

