% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{names,Spectra-method}
\alias{names,Spectra-method}
\title{The Names of a \code{Spectra} object}
\usage{
names(x)
}
\arguments{
\item{x}{a \code{Spectra} object}
}
\description{
Retrieve  the names of \code{Spectra} object
}
\examples{

x <- spc.example_spectra() 
names(x)

}
