% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{nrow,Spectra-method}
\alias{nrow,Spectra-method}
\title{The Number of rows  of a \code{Spectra} object}
\usage{
nrow(x)
}
\arguments{
\item{x}{a \code{Spectra} object}
}
\description{
\code{nrow} and \code{ncol} return the number of rows or columns present in a \code{Spectra} object
}
\examples{
x <- spc.example_spectra() 
ncol(x)  #501 
nrow(x)  #26


}
