% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcHeader-Methods.R
\docType{methods}
\name{show,BiooHeader-method}
\alias{show,BiooHeader-method}
\title{Show a bioo object}
\usage{
show(object)
}
\arguments{
\item{object}{of bioo}
}
\description{
Display  bioo object
}
\examples{
x=spc.example_spectra()
show(x)


}
\seealso{
\code{\link{show {methods}}}
}

