% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcHeader-Methods.R
\docType{methods}
\name{show,SpcHeader-method}
\alias{show,SpcHeader-method}
\title{Show a SpcHeader object}
\usage{
\S4method{show}{SpcHeader}(object)
}
\arguments{
\item{object}{of class SpcHeader}
}
\description{
Display a SpcHeader object
}
\examples{
x=spc.example_spectra()
show(x@header)

}
\seealso{
\code{\link{show}}
}
