% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\docType{methods}
\name{show,SpcList-method}
\alias{show,SpcList-method}
\title{Show a Spclist object}
\usage{
show(x)
# or 
x
}
\arguments{
\item{x}{a \code{Spclist} object}
}
\value{
show returns an invisible \code{NULL}
}
\description{
Display a \code{Spclist} object
}
\examples{
x <- spc.example_spectra()
BL = spc.makeSpcList(x,"STATION")
show(BL)


}

