% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{show,Spectra-method}
\alias{show,Spectra-method}
\title{Show a spectra object}
\usage{
show(x)
}
\arguments{
\item{x}{a spectral object}
}
\value{
show returns an invisible \code{NULL}
}
\description{
Display the objects, by printing,plotting and so on. 
This function will be invoked for automatic pirinting
}
\examples{
x <- spc.example_spectra()
show(x)



}

