% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\docType{methods}
\name{sort,list-method}
\alias{sort,list-method}
\title{Sorting   elements of a spclist}
\usage{
spc.data2header(x, which.col, decreasing = FALSE, ...)
}
\arguments{
\item{x}{\code{spclist} object}

\item{decreasing}{gical. Should the sort be increasing or decreasing}

\item{...}{arguments to be passed to or from methods}
}
\description{
Sort elements of a spclist  into ascending or descending order
}
\examples{
 
}

