% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{spc.bbox2lines}
\alias{spc.bbox2lines}
\alias{spc.bbox2lines,Spatial-method}
\alias{spc.bbox2lines,STI-method}
\alias{spc.bbox2lines,Spectra-method}
\title{Constructs a rectangle with a \code{Spectra} object}
\usage{
spc.bbox2lines(object)

\S4method{spc.bbox2lines}{Spatial}(object)

\S4method{spc.bbox2lines}{STI}(object)

\S4method{spc.bbox2lines}{Spectra}(object)
}
\arguments{
\item{object}{spectra object t}
}
\description{
Constructs a rectangle of sp::Lines using the bounding box of a \code{Spectra} object.
}
\examples{
sp=spc.example_spectra()
spc.bbox2lines(sp)

}
