% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.cname.construct}
\alias{spc.cname.construct}
\title{Generating column names for a spectra object}
\usage{
spc.cname.construct(object, value)
}
\arguments{
\item{object}{of Spectra}

\item{value}{ShortName}
}
\value{
vector of character
}
\description{
Function for a spectra object generates that  column names and it is made of  combination of @shortName  and @Wavelenght
}

