% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{spc.colMeans}
\alias{spc.colMeans}
\alias{spc.colMeans,Spectra-method}
\title{Computes the mean along the rows of a \code{Spectra} object}
\usage{
spc.colMeans(object)

\S4method{spc.colMeans}{Spectra}(object)
}
\arguments{
\item{object}{a \code{Spectra} object}
}
\description{
Computes the mean along the rows of a \code{Spectra} object. The method finds the measurement 
closest in time to the mean time and keeps the spatial/time attributes as well as Ancillary
data table (@data) associated to that measurement as that of the mean spectra
}
\examples{
sp=spc.example_spectra()
spc.colMeans(sp)
}
