% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\docType{methods}
\name{spc.data2header,list-method}
\alias{spc.data2header,list-method}
\title{Populate fields of header slot of a Spclist object  using data from data slot}
\usage{
spc.data2header(object,dataname,headerfield,compress)
}
\arguments{
\item{object}{\code{Spclist} object}

\item{dataname}{A character object specifying the name of data column to be used}

\item{headerfield}{A character object specifying the name of the header field to be changed}

\item{compress}{true or false}
}
\value{
object of class \code{Spclist}
}
\description{
Populates a field of header of a Spclist object with a column data from data slot.
}
\details{
This function extracts data from a column of the @data slot (specified by dataname)  
and creates a new header of a Spclist object field with it. If headerfield is not provided, the name 
of the new header field will be the same as dataname. 
The name of the new header field can be overwritten by providing headerfield.
If all the incoming data rows (dataname) are the same, information put into the header 
can be compressed by selecting compress=TRUE (default is FALSE). This would take only the first element 
from the @data column.
}
\examples{
sp=spc.example_spectra()
BL = spc.makeSpcList(sp,"STATION")
BL=spc.data2header(BL,"CRUISE")
spc.getheader(BL,"CRUISE")
spc.getheader(BL)
BL=spc.data2header(BL,"CAST","ProjectCast",compress = F)
spc.getheader(BL,"ProjectCast")

}

