% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.data2header}
\alias{spc.data2header}
\title{Set or cahenge data in header}
\usage{
spc.data2header(x,headerfield,dataname, compress )
}
\arguments{
\item{headerfield}{data column}

\item{dataname}{list spectra object}

\item{compress}{true or false}

\item{x}{spectra object}
}
\description{
Function  populates  the header of each element  with a column
}
\details{
The function  can be used to populate  the header of each element  with a column.
Since we are certain that all the columns of all individuals elements are the same,
we can compress the column (taking only the first element from data column into the header.
}

