% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\docType{methods}
\name{spc.getheader,list-method}
\alias{spc.getheader,list-method}
\title{Extract a field of the header slot of a \code{Spclist} object}
\usage{
spc.getheader(object,name)
}
\arguments{
\item{object}{A  \code{Spectra} object}

\item{name}{of the header field to be extracted}
}
\description{
Extracts the value of a field in the header slot of \code{Spclist} object
}
\examples{
sp=spc.example_spectra()
BL = spc.makeSpcList(sp,"STATION")
spc.getheader(BL)
#or
spc.getheader(BL,"Latitude")

}
\seealso{
\code{\link{spc.setheader}}
}

