% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.getheader}
\alias{spc.getheader}
\title{Extract the first or last part of column of named of a Spectra object}
\usage{
spc.getheader(x,name)
}
\arguments{
\item{name}{name of spactra object}

\item{x}{spectra object}
}
\description{
Function extracts the first or last part of column acording to mentioned name insade of  a Spectra object
}

