% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.getinvalid.idx}
\alias{spc.getinvalid.idx}
\title{Get the selected \code{Spectra} as invalid}
\usage{
spc.getinvalid.idx(object)
}
\arguments{
\item{object}{A \code{Spectra} object}
}
\description{
Extract the indexes stored as invaild variable
}

