% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{spc.getwavelengths,Spectra-method}
\alias{spc.getwavelengths,Spectra-method}
\title{@rdname spc.getwavelengths}
\usage{
\S4method{spc.getwavelengths}{Spectra}(object)
}
\description{
@rdname spc.getwavelengths
}
