% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.getwavelengths}
\alias{spc.getwavelengths}
\title{Getting wave lenghts in a \code{Spectra} object}
\usage{
spc.getwavelengths(object)
}
\arguments{
\item{object}{A \code{Spectra} object}
}
\description{
Get wave lenghts insade of  a \code{Spectra} object
}
\examples{
 x <- spc.example_spectra()
 spc.getwavelengths(x)

}

