% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.header.infos}
\alias{spc.header.infos}
\title{Getting as input the \code{Spectra} heade}
\usage{
spc.header.infos(header)
}
\arguments{
\item{header}{A \code{Spectra} header}
}
\description{
This internal function takes a \code{Spectral} header as a list for input and 
converts its elements to numbers (when possible) and
evaluates its elements in case the text contains some R code.
}
\examples{
sp=spc.example_spectra()
spc.header.infos(sp@header)

}
