% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.header2data}
\alias{spc.header2data}
\title{Get header for data}
\usage{
spc.data2header(object,headerfield,dataname )
}
\arguments{
\item{object}{A \code{Spectra} object}

\item{headerfield}{data column}

\item{dataname}{list \code{Spectra} object}

\item{...}{arguments to be passed to or from other methods}
}
\value{
object of class \code{Spectra}
}
\description{
Get  the header for data of each element  with a column
}
\details{
If header element has length >1, its type is checked. If it is "character",
its elements will be pasted using paste(...,collapse="|"). If it is another 
type, only the first element will be taken.
}

