% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.import.text}
\alias{spc.import.text}
\title{Importing  text into R}
\usage{
spc.import.text(...)
}
\arguments{
\item{...}{name of text}
}
\description{
Texts can be imported back into R with this function
}
\examples{
x=spc.example_spectra()
 spc.export.text(x,filename="anap.txt")
aa=spc.import.text("anap.txt")
spc.plot(aa)
}

