% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.interp.spectral}
\alias{spc.interp.spectral}
\title{Interpolate spectral values}
\usage{
spc.interp.spectral(source1,target_lbd,show.plot=FALSE)
}
\arguments{
\item{source1}{A \code{Spectra} object}

\item{target_lbd}{numeric vector giving desired wavelengths}

\item{show.plot}{logical TRUE if a graphical representation is required}
}
\description{
Estimate spectral data at a new set of wavelengths through interpolation
}
\examples{
sp=spc.example_spectra()
lbd = as.numeric(c(412,440,490,555,670))
spc.interp.spectral(sp[,lbd],c(430,450,500),show.plot=TRUE)

}

