% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\docType{methods}
\name{spc.invalid.detect,list-method}
\alias{spc.invalid.detect,list-method}
\title{Determinate invalid records insade of a \code{spclist} object}
\usage{
spc.invalid.detect(source1)
}
\arguments{
\item{source1}{A  \code{spclist} object}
}
\description{
Detect invalid records (rows) inside of a \code{spclist} object and returns logical object
}
\examples{
sp=spc.example_spectra()
BL = spc.makeSpcList(sp,"CAST")
invalid=spc.getheader(BL)
show(invalid)

}
