% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\name{spc.lapply}
\alias{spc.lapply}
\title{Apply a function over  a Spclist object}
\usage{
spc.lapply(X, FUN, ...)
}
\arguments{
\item{X}{A \code{Spectra} object}

\item{FUN}{The function to be applied to each element of \code{Spectra} object}

\item{...}{Optional arguments to FUN.}
}
\description{
spc.lapply returns a list of the same length as a \code{Spectra} object, each element of which is the result of applying FUN to the corresponding element of \code{Spectra} object.
}
\examples{
sp=spc.example_spectra() 
BL = spc.makeSpcList(sp,"STATION")
BL2=spc.lapply(BL, function(x) x=x+1)
class(BL2)
 BL2@by
 # if we use spc.lapply(), we keep @by slot but if we use lapply(), we lose it
 BL2=lapply(BL, function(x) x=x+1)
 class(BL2)
 
}

