% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.lines}
\alias{spc.lines}
\title{Add Connected Line Segments to a Plot}
\usage{
spc.lines(x,...)
}
\arguments{
\item{x}{a \code{Spectra} vector of points to join plot}
}
\description{
Taking coordinates given in various ways and joining the corresponding points with line segments.
}

