% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.makeSpcList}
\alias{spc.makeSpcList}
\title{Conversion from \code{Spectra} to \code{Spclist}}
\usage{
spc.makeSpcList(myobj, name)
}
\description{
Conversion from \code{Spectra} to \code{Spclist} using a data field
 @usage 
 spc.makeSpcList(myobj, name)
 @param myobj a \code{Spectra} object
 @param name  name of station of a \code{Spectra} object
}
\examples{
sp <- spc.example_spectra()
BL = spc.makeSpcList(sp,"CAST")
show(BL)
}
