% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\name{spc.plot.depth.overlay}
\alias{spc.plot.depth.overlay}
\title{Plotting \code{SpcList} object}
\usage{
spc.plot(object, X, lab_cex, ...)
}
\arguments{
\item{object}{a \code{SpcList} data}
}
\description{
Generating plot of the contents of a \code{SpcList} object overlay with respect to depth
}
\examples{
x <- spc.example_spectra()
BL = spc.makeSpcList(sp,"CAST")
spc.plot.depth.overlay(BL)


}

