% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{spc.plot.depth.plotly,Spectra-method}
\alias{spc.plot.depth.plotly,Spectra-method}
\title{Display a Spectra object}
\usage{
\S4method{spc.plot.depth.plotly}{Spectra}(sp, column, plot.max = 10)
}
\arguments{
\item{sp}{A \code{Spectra} object}

\item{column}{Number or name , defoult value is 10 if a number or name has not been entered}
}
\description{
Plot a \code{Spectra} object with respect to depth
}
\examples{
BL = spc.makeSpcList(sp,"CAST")
p1<-spc.plot.plotly.depth.test(BL[[5]])
p1<-layout(title=paste("CAST =", BL[[5]]$CAST[1]))
p2<-spc.plot.plotly.depth.test(BL[[4]])
p2<-layout(title=paste("CAST =", BL[[4]]$CAST[1]))
p <- subplot(p1, p2,  margin = 0.05)
p <- layout(showlegend = FALSE)
p
}

