% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\name{spc.plot.grid}
\alias{spc.plot.grid}
\title{Plotting \code{SpcList} object overlay}
\usage{
spc.plot.grid(x,FUN, nnrow, nncol, mar=c(4,4.5,1,0.5),oma = c(0,0,0,0), lab_cex, ...)
}
\arguments{
\item{x}{a \code{SpcList} data}
}
\description{
Generating plot of the contents of a \code{SpcList} object respect to grid
}
\examples{
x <- spc.example_spectra()
BL = spc.makeSpcList(sp,"CAST")
spc.plot.grid(BL)

}

