% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.plot.map.plotly}
\alias{spc.plot.map.plotly}
\title{Display a Spectra object}
\usage{
spc.plot.map.plotly(sp, hover_field = "row", color = "#FF0000",
  opacity = 1)
}
\arguments{
\item{sp}{A \code{Spectra} object}

\item{hover_field}{A character, column  names of sp object to be used  in hover box}

\item{color}{Determine color of points}

\item{opacity}{The opacity transparency of the glyph 
between 0 (transparent) and 1 (opaque)}
}
\description{
Create a point map with ploty engine using \code{Spectra} rows
}
\examples{
sp <- spc.example_spectra()
spc.plot.map.plotly(sp)

}
