% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\name{spc.plot.overlay}
\alias{spc.plot.overlay}
\title{Plotting multiple  \code{Spectra} objects insade a \code{SpcList}}
\usage{
spc.plot(x,lab_cex=1,leg_idx=T, type="l", lty=1,lwd=1, col, ...)
}
\arguments{
\item{...}{any further arguments of plot}

\item{x}{a \code{SpcList} data}

\item{lab_cex}{vector of character expansion sizes, used cyclically}

\item{type}{character string (length 1 vector) or vector of 1-character strings indicating the type of plot for each column of y,}

\item{lwd}{vector of line widths}

\item{lty}{vector of line types}

\item{col}{columns of a \code{SpcList} data}

\item{leg_idx}{logical, to display legend of index or not, default is TRUE}
}
\description{
Generating plot of the contents of a \code{SpcList} object overlay
}
\examples{
sp <- spc.example_spectra()
BL = spc.makeSpcList(sp,"CAST")
spc.plot.overlay(BL)


}

