% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.plot}
\alias{spc.plot}
\title{Plotting \code{Spectra} object}
\usage{
spc.plot(x,...)
}
\arguments{
\item{x}{a \code{Spectra} data}
}
\description{
Generating plot of the contents of a \code{Spectra} object
}
\examples{
x <- spc.example_spectra()
spc.plot(x)


}

