% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.plot.time.plotly}
\alias{spc.plot.time.plotly}
\title{Display a Spectra object}
\usage{
spc.plot.time.plotly(sp, column, plot.max = 10)
}
\arguments{
\item{sp}{A \code{Spectra} object}

\item{column}{Number or name , defoult value is 10 if a number or name has not been entered}
}
\description{
Plot a \code{Spectra} object with respect to time
}
\examples{
spc.plot.time.plotly(sp)
spc.plot.time.plotly(sp, plot.max = 3)
spc.plot.time.plotly(sp, c("anap_450","anap_550","anap_650"))
}

