% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{spc.plot.time}
\alias{spc.plot.time}
\alias{spc.plot.time,Spectra-method}
\title{Plotting \code{Spectra} object}
\usage{
spc.plot.time(object, ...)

\S4method{spc.plot.time}{Spectra}(object, Y, maxSp = 50, xdata = "time",
  lab_cex, lwd = 2, ...)
}
\arguments{
\item{object}{a \code{Spectra} data}

\item{...}{any further arguments of plot}
}
\description{
Generating plot of the contents of a \code{Spectra} object with respect to time
}
\examples{
x <- spc.example_spectra()
spc.plot.time(x)
}
\seealso{
\code{\link{spc.plot.depth}}
}

