% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{spc.rbind,STIDF-method}
\alias{spc.rbind,STIDF-method}
\title{Combine \code{STIDF} objects by Rows}
\usage{
\S4method{spc.rbind}{STIDF}(...)
}
\arguments{
\item{...}{\code{STIDF} object}
}
\description{
Take a \code{STIDF} objects and combine by rows
}
\examples{
x <- spc.example_spectra()
nrow(x)  #[1] 26
x2 <- spc.rbind(as(x, "STIDF"),as(x, "STIDF"))
nrow(x2)  #[1] 52

}
