% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcHeader-Methods.R
\name{spc.ScanFileName}
\alias{spc.ScanFileName}
\title{Scan File Name}
\usage{
spc.ScanFileName (template,in.listnames,Scan)
}
\arguments{
\item{object}{of class SpcHeader}
}
\description{
Scan name of file
}
\examples{
sp=spc.example_spectra()


}

