% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\docType{methods}
\name{spc.setheader<-,list,list-method}
\alias{spc.setheader<-,list,list-method}
\title{Set a field of the header slot of a Spclist object}
\usage{
spc.setheader(object)<-value
}
\arguments{
\item{object}{A \code{Spclist} object}

\item{value}{A list of SpcHeader objects to be set}
}
\description{
This function sets or changes the header slot of each \code{Spectra} object within 
\code{Spclist} object with provided list of SpcHeader objects. Both \code{object}
and \code{value} must be of same lenght.
}
\examples{
sp=spc.example_spectra()
BL = spc.makeSpcList(sp,"STATION")
a=new("SpcHeader") # create new SpcHeader object
#Make 4 copies of the SpcHeader object, with different 'Station' fields
new_headers = lapply(1:length(a), function(x) a)
new_headers = lapply(1:length(new_headers), function(x) {new_headers[[x]]$Station=paste0("Station",x); new_headers[[x]]})
new_headers
#Replace the 'Station' field of each Spectra object in the SpcList
spc.setheader(BL,"Station") <- new_headers
spc.getheader(BL,"Station")
}
\seealso{
\code{\link{spc.getheader}}
}

