% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\docType{methods}
\name{spc.setheader<-,list-method}
\alias{spc.setheader<-,list-method}
\title{Set a field of the @header slot of a \code{SpcList} object}
\usage{
\S4method{spc.setheader}{list}(object) <- value
}
\arguments{
<<<<<<< HEAD
\item{value}{Object of class SpcList}

\item{x}{A \code{SpcList} object}
=======
\item{object}{A \code{SpcList} object.}
>>>>>>> f1e6acacecbb765302786737b78406f6151356f1

\item{value}{Object of class SpcList.}

<<<<<<< HEAD
\item{...}{arguments to be passed to or from other methods}
=======
\item{...}{arguments to be passed to or from other methods.}
>>>>>>> f1e6acacecbb765302786737b78406f6151356f1
}
\description{
Function sets or changes the value of a field in the header slot of \code{SpcList} object.
}
\examples{
sp=spc.example_spectra()
BL=spc.makeSpcList(sp,"CAST")
a=new("SpcHeader") # create new SpcHeader class
a$Longitude=123 
spc.setheader(BL[[1]],"Station") <- a
h=spc.getheader(BL[[1]])
h

}
