% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{spc.setheader<-}
\alias{spc.setheader<-}
\alias{spc.setheader<-,Spectra-method}
\title{Set a field of the @header slot of a \code{Spectra} object}
\usage{
<<<<<<< HEAD
<<<<<<< HEAD:man/spc.setheader.Rd
spc.setheader(x,name,...)<-value

\S4method{spc.setheader}{Spectra}(object) <- value
=======
spc.setheader(object,name,value,...)
>>>>>>> khaliilhaji-dev:man/spc.setheader-set.Rd
=======
spc.setheader(object) <- value

\S4method{spc.setheader}{Spectra}(object) <- value
>>>>>>> khaliilhaji-devv
}
\arguments{
\item{object}{A \code{Spectra} object}

<<<<<<< HEAD
<<<<<<< HEAD
\item{x}{A \code{Spectra} object}

\item{name}{of the header field to be setted}

=======
\item{name}{of the header field to be setted}

\item{value}{Object of class SpcHeader}

>>>>>>> 83febb7f733cd671c07061f9ea3eb580475924bc
\item{...}{arguments to be passed to or from other methods}
=======
\item{value}{Object of class SpcHeader}

\item{name}{of the header field to be setted}
>>>>>>> khaliilhaji-devv
}
\description{
Function sets or changes the value of a field in the header slot of \code{Spectra} object
}
\examples{
sp=spc.example_spectra()
a=new("SpcHeader") # create new SpcHeader class
a$Longitude=123 
spc.setheader(sp,"Station") <- a
sp@header

}
\seealso{
\code{\link{spc.getheader}}
}

