% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.setinvalid.idx<-}
\alias{spc.setinvalid.idx<-}
\title{Set the selected \code{Spectra} as invalid}
\usage{
spc.setinvalid.idx(object,value)
}
\arguments{
\item{value}{Numeric}

\item{x}{A \code{Spectra} object}
}
\description{
Stores the indexes as invaild variable
}

