% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.setselected.idx<-}
\alias{spc.setselected.idx<-}
\title{Set index to a spectra object}
\usage{
spc.setselected.idx(x,value)
}
\arguments{
\item{value}{index for spectra object}

\item{x}{spectra object}
}
\description{
Function set or change  index inside of  a spectra object and show it
}

