% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.setwavelengths<-}
\alias{spc.setwavelengths<-}
\title{Setting wavelengths  in a \code{Spectra} object}
\usage{
spc.setwavelengths(object) <- value
}
\arguments{
\item{object}{A \code{Spectra} object}

\item{value}{Numeric}
}
\description{
Function  to change or set wavelengths  inside  of  a \code{Spectra} object
}
\examples{
 x <- spc.example_spectra()
 show(x)
 spc.setwavelengths(x) <- 300:800
 show(x)
<<<<<<< HEAD
 
=======
>>>>>>> 14134e3400d5c89dbe16674b95ec8b8c419d95e3
}
\seealso{
\code{\link{spc.getwavelengths}}
}

