% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.STI.stdistance}
\alias{spc.STI.stdistance}
\title{Report the space and time distance of each row of the STI-inherited object}
\usage{
spc.STI.stdistance(...,searched)
}
\arguments{
\item{searched}{a data.frame}

\item{...}{STI-inherited object master}
}
\description{
Function for a spectra object reports the space and time distance of each row of the STI-inherited object
searched to the corresponding row of the STI-inherited object master
}
\details{
Reports the space and time distance of each row of the STI-inherited object
searched to the corresponding row of the STI-inherited object master. Outputs 
a data.frame, with two columns : time2master ("difftime", in seconds) and 
distance2master ("numeric", in meters)
}

