% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\docType{methods}
\name{spc.updateheader<-,list-method}
\alias{spc.updateheader<-,list-method}
\title{Update a field of the @header slot of a \code{spclist} object}
\usage{
\S4method{spc.updateheader}{list}(object, Name) <- value
}
\arguments{
\item{object}{A \code{Spectra} objec}

\item{...}{arguments to be passed to or from other methods}

\item{name}{of the header field to be updated}
}
\description{
Updates or changes the value of a field in the header slot of \code{spclist} object
}
\examples{
sp=spc.example_spectra()
BL=spc.makeSpcList(sp,"CAST")
BL[[1]]@header
spc.updateheader(BL[[1]],"Station")<-11
BL[[1]]@header

}

