% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcHeader-Class.R
\docType{class}
\name{SpcHeader-class}
\alias{SpcHeader-class}
\title{Constructor for the \code{SpcHeader} class a \code{SpcHeader} class object}
\arguments{
\item{object}{of class SpcHeader}

\item{...}{a additional variable}
}
\description{
Creates an empty header object initilazed with defult values
}
\examples{
new("SpcHeader")
}

