% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcHeader-Methods.R
\name{SpcHeaderAdd}
\alias{SpcHeaderAdd}
\title{Set a field of the @header slot of a \code{SpcHeader} class object}
\usage{
SpcHeaderAdd (object,Name,Value)
}
\arguments{
\item{object}{of class SpcHeader}

\item{Name}{a character variable}

\item{Value}{a numeric variable}
}
\description{
Function add the value of a field in the header slot of \code{SpcHeader} class object
}
\examples{
sp=spc.example_spectra()
sp@header
sp@ShortName
sp@header=SpcHeaderAdd(sp@header,sp@ShortName,10)
sp@header
}
