% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcHeader-Class.R
\docType{class}
\name{SpcHeaderList-class}
\alias{SpcHeaderList-class}
\title{\code{SpcHeaderList} class.}
\description{
Definition for \code{SpcHeaderList}. This class provides
a collection of multiple \code{SpcHeader} objects inside a list.
}
\examples{
h1 = new("SpcHeader")
h2 = new("SpcHeader")
as(list(h1, h2), "SpcHeaderList")
new("SpcHeaderList") 
}
