% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Class.R
\docType{class}
\name{SpcList-class}
\alias{SpcList-class}
\title{SpcList class definition}
\description{
Definition for \code{SpcList}, a class to store multiple \code{Spectra}
objects inside a list-like object. See the help of the constructor function \code{\link{SpcList}}.
}
\section{Slots}{

\describe{
\item{\code{by}}{character, Determines the header field in the Spectra objects
within the SpcList that describes how they are different one from the other.}
}}

