% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\name{SpcList}
\alias{SpcList}
\title{Construction a Spclist object}
\usage{
SpcList(listofsp)
}
\arguments{
\item{listofsp}{A \code{Spclist} object}
}
\description{
Create a \code{Spclist} object by using list of \code{Spactra} object
}
\examples{
sp$CLASS = factor(c(rep("c",12), rep("a",6), rep("b",8)))
sp$CLASS 
SpcList(sp$CLASS)
}

