% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{$}
\alias{$}
\alias{$,Spectra}
\alias{$}
\alias{$<-,Spectra}
\title{Extract or replace parts of a \code{Spectra} object}
\usage{
x[i] 
x[i, j] 
x$i

\S4method{$}{Spectra}(x, name) <- value
}
\arguments{
\item{\code{Spectra}}{object from which to extract element(s) or in which to replace element(s)}

\item{i}{A numeric (row index) variable}

\item{j}{A character (column name) or a numeric (column index) variable}
}
\description{
Operators acting on \code{Spectra} objects to extract parts

Operators acting on \code{Spectra} objects to replace parts
}
\examples{
 sp<-spc.example_spectra()
 # spc.colnames() is used to extract column names
 head(spc.colnames(sp))
 head(sp$anap_300)
 sp[,"anap_345"] #returns Spectra object with only one channel (column)
 sp[1:3,"anap_345"] #returns Spectra object with first 3 rows and only one channel (column)
 
 # spc.colnames() is used to extract column names
 head(spc.colnames(sp))
 head(sp$anap_300)
 sp[,"anap_345"]
}
