% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{[,Spectra-method}
\alias{[,Spectra-method}
\title{Extract or replace parts of a spectra object}
\usage{
x[i]
}
\description{
Operators acting on spectra object and spectra lists to extract or replace parts.
}
\details{
These operators are generic. You can write methods to handle indexing of specific classes of objects
}

