% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{subset,Spectra-method}
\alias{subset,Spectra-method}
\title{Subsetting for spectra and spcList classes}
\usage{
subset(x,y,select)
}
\arguments{
\item{x}{spectra object}

\item{select}{condition selected}

\item{y}{subset}
}
\description{
Subsetting can be achieved using the implementation of the R function subset() for Spectra and SpcList classes
It is possible to perform a row-wise selection
}

