% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\docType{methods}
\name{$,SpcList-method}
\alias{$,SpcList-method}
\title{Extract or replace parts of a \code{SpcList} object}
\usage{
x[i] 
x[i, j] 
x[[i]] 
x$i
}
\arguments{
\item{\code{Spectra}}{object from which to extract element(s) or in which to replace element(s)}

\item{i}{A numeric (row index) variable}

\item{j}{A character (column name) or a numeric (column index) variable}
}
\description{
Operators acting on  \code{Spectra} objects  to extract or replace parts
}
\examples{
  sp<-spc.example_spectra()
  BL = spc.makeSpcList(sp,"STATION")
  
  #Extract station 394 (returns Spectra object)
  BL$`394`
  
  BL@by="CRUISE"
  BL[[1]]$CRUISE="Cruise1"
  BL[[2]]$CRUISE="Cruise2"
  BL[[3]]$CRUISE="Cruise3"
  BL[[4]]$CRUISE="Cruise4"
  names(BL)
  BL$Cruise4

}
