% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\docType{methods}
\name{$,SpcList-method}
\alias{$,SpcList-method}
\title{Extract or replace parts of a Spclist object}
\usage{
x[i] 
x[i, j] 
x[[i]] 
x$i #More usage cases to be added
}
\arguments{
\item{\code{Spclist}}{object from which to extract element(s) or in which to replace element(s)}

\item{i}{A numeric (row index) variable}

\item{j}{A character (column name) or a numeric (column index) variable}
}
\description{
Operators acting on  \code{Spclist} objects  to extract or replace parts
}
\examples{
 sp<-spc.example_spectra()
 BL = spc.makeSpcList(sp,"STATION")
 # names() is used to show that a_nap ,  
 names(BL)
 BL$a_nap
 BL["a_nap"]
 

}

