% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{$,Spectra-method}
\alias{$,Spectra-method}
\title{Extract or replace parts of a spectra object}
\usage{
x[i] <- spc.example_spectra()
x[i, j, ...] <- spc.example_spectra()
x[[i]] <- spc.example_spectra()
x$i <- spc.example_spectra()
}
\arguments{
\item{Spectra}{object from which to extract element(s) or in which to replace element(s)}
}
\description{
Operators acting on  spectral objects such as vectors, matrices, arrays and lists to extract or replace parts
}
\examples{
 
 


}

