% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{dim,Spectra-method}
\alias{dim,Spectra-method}
\title{Dimensions of a \code{Spectra} object.}
\usage{
\S4method{dim}{Spectra}(x)
}
\arguments{
\item{x}{a \code{Spectra} objectlong-format \code{data.frame}}
}
\value{
Returns a numeric vector containing \code{nrow} and \code{ncol} of the \code{Spectra} object.
}
\description{
\code{Spectra} Creates an instance of class \code{Spectra}.
}
\details{
This
}
\examples{
fnm = file.path(base::system.file(package = "Spectral"),"test_data","particulate_absorption.csv.gz")
}

