% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{head,Spectra-method}
\alias{head,Spectra-method}
\title{Return the first or last part of a \code{Spectra} object}
\usage{
\S4method{head}{Spectra}(x, ...)
}
\arguments{
\item{x}{a \code{Spectra} object}

\item{...}{arguments to be passed to or from other methods}
}
\value{
Returns a matrix (\code{Spectra} data)
}
\description{
Return the first or last parts of a \code{Spectra} object
}
\examples{
x <- spc.example_spectra()
head(x)

}
