% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\docType{methods}
\name{names,SpcList-method}
\alias{names,SpcList-method}
\title{names of \code{SpcList} object}
\usage{
spc.colnames(x)
}
\arguments{
\item{x}{A \code{SpcList} object}
}
\value{
Returns the coulmn names of an object of class \code{SpcList} as a charecter vector.
}
\description{
Retrieve   names of a \code{SpcList} object
}
\examples{
sp <- spc.example_spectra()
BL = spc.makeSpcList(sp,"CAST")
names(BL)

}
