% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{names,Spectra-method}
\alias{names,Spectra-method}
\title{The Names of a Spectra object}
\usage{
names(x)
}
\arguments{
\item{x}{a Spectra object}
}
\description{
Function to get  the names of a Spectra object
}
\examples{

x <- spc.example_spectra() 
names(x)



}

