% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{ncol,Spectra-method}
\alias{ncol,Spectra-method}
\title{The Number of Columns  of a Spectra object}
\usage{
\S4method{ncol}{Spectra}(x)
}
\arguments{
\item{x}{A \code{Spectra} object}
}
\description{
\code{nrow} and \code{ncol} return the number of rows or columns of a \code{Spectra} object
}
\examples{
x <- spc.example_spectra() 
ncol(x)  #501 
nrow(x)  #26

}
