% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcHeader-Methods.R
\docType{methods}
\name{show,BiooHeader-method}
\alias{show,BiooHeader-method}
\title{Show a spectra object}
\usage{
show(object)
}
\arguments{
\item{object}{of spectra}
}
\description{
Display the object, by printing, plotting
}
\examples{
x=spc.example_spectra()
show(x)


}
\seealso{
http://127.0.0.1:41556/help/library/methods/help/showMethods
}

