% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\docType{methods}
\name{show,SpcList-method}
\alias{show,SpcList-method}
\title{Show a \code{SpcList} object}
\usage{
\S4method{show}{SpcList}(object)
}
\arguments{
\item{x}{a \code{SpcList} object}
}
\value{
show returns an invisible \code{NULL}
}
\description{
Display a \code{SpcList} object
}
\examples{
x <- spc.example_spectra()
BL = spc.makeSpcList(x,"CAST")
show(BL)
}
