% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{show,Spectra-method}
\alias{show,Spectra-method}
\title{Show a \code{Spectra} object}
\usage{
\S4method{show}{Spectra}(object)
}
\arguments{
\item{x}{a \code{Spectra} object}
}
\value{
show returns an invisible \code{NULL}
}
\description{
Display a \code{Spectra} object
}
\examples{
x <- spc.example_spectra()
show(x)



}
