% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\docType{methods}
\name{sort,list-method}
\alias{sort,list-method}
\title{Sort a Spclist object}
\usage{
\S4method{sort}{list}(x, decreasing = FALSE, which.col, ...)
}
\arguments{
\item{x}{A \code{Spclist} object}

\item{which.col}{A number or name of column and default is that decreaing is False}
}
\value{
Returns a \code{Spclist} object
}
\description{
Sort \code{Spclist} objects into ascending or descending order.
}
\examples{
sp=spc.example_spectra()
BL = spc.makeSpcList(sp,"STATION")
p=sort(BL,"anap_301", decreasing = TRUE)
p[[2]]$anap_301
#or 
p=sort(BL,"anap_301", decreasing = FALSE)
p[[2]]$anap_301
}

