% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{spc.cname.construct}
\alias{spc.cname.construct}
\alias{spc.cname.construct,Spectra-method}
\title{Generating column names for a \code{Spectra} object}
\usage{
spc.cname.construct(object, value)

\S4method{spc.cname.construct}{Spectra}(object, value)
}
\arguments{
\item{object}{A variable of class \code{Spectra}}

\item{value}{A \code{character}}
}
\value{
vector composed of characters
}
\description{
Generates column names taken from the @shortName and @Wavelenght slots of a \code{spectra} object. If \code{value} is 
omitted, the @ShortName slot is used by default.
}
\examples{
sp <- spc.example_spectra()
spc.cname.construct(sp)
spc.cname.construct(sp,"Newvar")

}
