% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.cname.construct}
\alias{spc.cname.construct}
\title{Generating column names for a spectra object}
\usage{
spc.cname.construct(...)
}
\arguments{
\item{...}{spectra object}
}
\description{
Function for a spectra object generates column names and it is combination of shortName  and Wavelenght
}

