% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.colnames}
\alias{spc.colnames}
\title{Column names of \code{Spectra} object}
\usage{
spc.colnames(x,...)
}
\arguments{
\item{x}{A \code{Spectra} object}

\item{...}{arguments to be passed to or from other methods}
}
\value{
Returns the coulmn names of an object of class \code{Spectra} as a charecter vector.
}
\description{
Retrieve  column names of a \code{Spectra} object
}
\examples{
x <- spc.example_spectra()
spc.colnames(x)
# or 
spc.colnames(x) <-spc.cname.construct(x)

}
\seealso{
\code{\link{spc.cname.construct}}
}

