% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.colnames}
\alias{spc.colnames}
\title{Column names}
\usage{
spc.colnames(x)
}
\arguments{
\item{x}{a matrix-like spectra object, with at least two dimensions for colnames.}
}
\value{
Returns the names of an object of class \code{Spectra}.
}
\description{
Retrieve  column names of a matrix-like spectra object.
}
\examples{
x <- spc.example_spectra()
spc.colnames(x)


}

