% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{spc.colnames}
\alias{spc.colnames}
\alias{spc.colnames,Spectra-method}
\title{Column names of \code{Spectra} object}
\usage{
spc.colnames(x)

\S4method{spc.colnames}{Spectra}(x)
}
\arguments{
\item{x}{A \code{Spectra} object}
}
\value{
Returns the column  names of an object of class \code{Spectra} as a character  vector.
}
\description{
Retrieve  column names of a \code{Spectra} object
}
\examples{
x <- spc.example_spectra()
spc.colnames(x)
# or 
spc.colnames(x) <- spc.cname.construct(x)
spc.colnames(x)
}
\seealso{
\code{\link{spc.cname.construct}}
}
