% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.data2header}
\alias{spc.data2header}
\title{Populate fields of header slot using data from data slot}
\usage{
spc.data2header(object,dataname,headerfield,compress,...)
}
\arguments{
\item{object}{\code{Spectra} object}

\item{dataname}{A character object specifying the name of @data column to be used}

\item{headerfield}{A character object specifying the name of the @header field to be changed
@param ... arguments to be passed to or from other methods}

\item{compress}{true or false}
}
\value{
object of class \code{Spectra}
}
\description{
Populates a field of @header with a column data from @data slot.
}
\details{
This function extracts data from a column of the @data slot (specified by dataname)  
and creates a new @header field with it. Ifa header field is not provided, the name 
of the new header field will be the same as dataname. 
The name of the new header field can be overwritten by providing header field.
If all the incoming data rows (dataname) are the same, information put into the header 
can be compressed by selecting compress=TRUE (default is FALSE). This would take only the first element 
from the @data column.
}
\examples{
sp=spc.example_spectra()
sp=spc.data2header(sp,"CAST")
sp@header
sp=spc.data2header(sp,"CAST","ProjectCast")
sp@header
sp$CAST=rep(33, nrow(sp))
sp=spc.data2header(sp,"CAST","ProjectCast", compress=TRUE)
sp@header
}
