% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.example_spectra}
\alias{spc.example_spectra}
\title{Create example of Spectral object}
\usage{
spc.example_spectra()
}
\description{
Example of Spectral object is created by the function
}
\examples{
sp = spc.example_spectra()
class(sp)
show(sp)

}
