% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.export.text}
\alias{spc.export.text}
\title{Exporting into text format}
\usage{
spc.export.text(x,filename)
}
\arguments{
\item{filename}{Name of the output text file}

\item{x}{A \code{Spectra} object}
}
\description{
Save the \code{Spectra} object on disk in text format
}
\examples{
x=spc.example_spectra()
spc.export.text(x,filename="anap.txt")
aa=spc.import.text("anap.txt")
dev.new()
spc.plot(aa)
}

