% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.export.text}
\alias{spc.export.text}
\title{Exporting into text format}
\usage{
spc.export.text(x,filename)
}
\arguments{
\item{filename}{text name}

\item{x}{a spectra object}
}
\description{
Save the spectra object on disk and Text files can be  saved with  this function
}
\examples{
x=spc.example_spectra()
spc.export.text(x,filename="anap.txt")
aa=spc.import.text("anap.txt")
dev.new()
spc.plot(aa)
}

