% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\docType{methods}
\name{spc.getheader,list-method}
\alias{spc.getheader,list-method}
\title{Extract a field of the @header slot of a \code{spclist} object}
\usage{
spc.getheader(x,name)
}
\arguments{
\item{name}{of the header field to be extracted}

\item{x}{A  \code{spclist} object}
}
\description{
Extracts the value of a field in the header slot of \code{spclist} object
}
\examples{
sp=spc.example_spectra()
BL = spc.makeSpcList(sp,"CAST")
BL[[1]]@header
spc.getheader(BL,"CAST")


}
