% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\docType{methods}
\name{spc.getheader,list-method}
\alias{spc.getheader,list-method}
\title{Extract a field of the @header slot of a \code{SpcList} object}
\usage{
\S4method{spc.getheader}{list}(object, name)
}
\arguments{
\item{object}{A  \code{SpcList} object}

\item{name}{of the header field to be extracted}
}
\description{
Extracts the value of a field in the header slot of \code{SpcList} object
}
\examples{
sp=spc.example_spectra()
BL = spc.makeSpcList(sp,"CAST")
BL[[1]]@header
spc.getheader(BL,"CAST")

}
