% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.getheader}
\alias{spc.getheader}
\title{Extract a field of the @header slot of a \code{Spectra} object}
\usage{
spc.getheader(x,name)
}
\arguments{
\item{name}{of the header field to be extracted}

\item{x}{A  \code{Spectra} object}
}
\description{
Extracts the value of a field in the header slot of \code{Spectra} object
}
\examples{
sp=spc.example_spectra()
sp@header
spc.getheader(sp,"Latitude")


}
\seealso{
\code{\link{spc.setheader<-}}
}
