% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.getheader}
\alias{spc.getheader}
\title{Extract a field of the @header slot of a spectra object}
\usage{
spc.getheader(x,name)
}
\arguments{
\item{name}{of the header field to be extracted}

\item{x}{spectra object}
}
\description{
Function extracts the value of a field in the header slot of spectra object
}
\examples{
sp=spc.example_spectra()
sp@header
spc.getheader(sp,"Latitude")


}

