% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.getinvalid.idx}
\alias{spc.getinvalid.idx}
\title{Get index of \code{Spectra} rows marked as invalid}
\usage{
spc.getinvalid.idx(object)
}
\arguments{
\item{object}{A \code{Spectra} object}
}
\value{
Logical vector
}
\description{
Extract the row indexes stored as invalid
}
\examples{
sp= spc.example_spectra()
spc.getinvalid.idx(sp) #No invalid rows

}
