% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.getinvalid.idx}
\alias{spc.getinvalid.idx}
\title{Get the selected spectra as invalid}
\usage{
spc.getinvalid.idx(x)
}
\arguments{
\item{x}{spectra object}
}
\description{
Function  extract the indexes stored as invaild variable
}

