% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.getselected.idx}
\alias{spc.getselected.idx}
\title{Extract index inside of a spectra object}
\usage{
spc.getselected.idx(x)
}
\arguments{
\item{x}{spectra object}
}
\description{
Function extracts index selected insade of a spectra object and show it
}

