% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.getwavelengths}
\alias{spc.getwavelengths}
\title{Extract wave lenghts of a \code{Spectra} object}
\usage{
spc.getwavelengths(object)
}
\arguments{
\item{object}{A \code{Spectra} object}
}
\value{
numeric vector of  wave lengths
}
\description{
Get wave lenghts inside of  a \code{Spectra} object
}
\examples{
 x <- spc.example_spectra()
 spc.getwavelengths(x)
<<<<<<< HEAD
 
=======
>>>>>>> 14134e3400d5c89dbe16674b95ec8b8c419d95e3
}
\seealso{
\code{\link{spc.setwavelengths<-}}
}

