% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.getwavelengths}
\alias{spc.getwavelengths}
\title{Getting wave lenghts in a spectral object}
\usage{
spc.getwavelengths(...)
}
\arguments{
\item{...}{spectra object}
}
\description{
Function get wave lenghts insade of  a spectral object
}
\examples{
 x <- spc.example_spectra()
 spc.getwavelengths(x)

}

