% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R, R/SpcList-Methods.R
\docType{methods}
\name{spc.header2data}
\alias{spc.header2data}
\alias{spc.header2data}
\alias{spc.header2data}
\title{Get header for data}
\usage{
spc.header2data(object,headerfield,dataname,...)

spc.header2data(object,headerfield,dataname,compress=TRUE,...)

\S4method{spc.header2data}{SpcList}(object, headerfield, dataname,
  compress = TRUE, ...)
}
\arguments{
\item{object}{A \code{Spectra} object}

\item{headerfield}{data column}

\item{dataname}{list \code{Spectra} object}

\item{...}{arguments to be passed to or from other methods}

\item{dataname}{list \code{spclist} object}

\item{object}{A \code{spclist} object}

\item{headerfield}{data column}
}
\value{
object of class \code{Spectra}

object of class \code{spclist}
}
\description{
Get  the header for data of each element  with a column

Get  the header for data of each element  with a column
}
\details{
If header element has length >1, its type is checked. If it is "character",
its elements will be pasted using paste(...,collapse="|"). If it is another 
type, only the first element will be taken.

If header element has length >1, its type is checked. If it is "character",
its elements will be pasted using paste(...,collapse="|"). If it is another 
type, only the first element will be taken.
}
\examples{
sp <- spc.example_spectra()
spc.updateheader(sp, "Zone")<- "ZoneA"
sp <- spc.header2data(sp, "Zone")
sp$Zone




sp <- spc.example_spectra()
BL=spc.makeSpcList(sp,"CAST")
spc.updateheader(BL[[1]], "Zone")<- "Zone"
BL[[1]] <- spc.header2data(BL[[1]], "Zone")
BL[[1]]$Zone
}
