% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.import.text}
\alias{spc.import.text}
\title{Importing  spectral data from a text file}
\usage{
spc.import.text(filename, sep,...)
}
\arguments{
\item{filename}{Name of input text file}

\item{sep}{Field delimiter to be used}

\item{...}{arguments to be passed to or from other methods.}
}
\description{
Spectral data already saved by spc.export.text() can be imported back 
as a \code{Spectral} object.
}
\examples{
x=spc.example_spectra()
spc.export.text(x,filename="anap.txt")
aa=spc.import.text("anap.txt")
spc.plot(aa)
}
\seealso{
\code{\link{spc.export.text}}
}
