% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.invalid.detect}
\alias{spc.invalid.detect}
\title{Determinate invalid things insade of a \code{Spectra} object}
\usage{
spc.invalid.detect(x)
}
\arguments{
\item{x}{A  \code{Spectra} object}
}
\description{
Detect invalid things insade of a \code{Spectra} object and returns logical object
}
\examples{
sp=spc.example_spectra()
invalid=spc.invalid.detect(sp)
show(invalid)
}
