% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{spc.invalid.detect}
\alias{spc.invalid.detect}
\alias{spc.invalid.detect,Spectra-method}
\title{Determinate invalid things insade of a \code{Spectra} object}
\usage{
spc.invalid.detect(source1)

\S4method{spc.invalid.detect}{Spectra}(source1)
}
\arguments{
\item{source1}{A  \code{Spectra} object}
}
\description{
Detect invalid things insade of a \code{Spectra} object and returns logical object
}
\examples{
sp=spc.example_spectra()
invalid=spc.invalid.detect(sp)
show(invalid)

}
