% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\name{spc.lapply}
\alias{spc.lapply}
\title{Apply a function over a Spclist}
\usage{
spc.lapply(X, FUN, ...)
}
\arguments{
\item{X}{A \code{spclist} object}

\item{...}{optional arguments to FUN.}
}
\value{
FUN  function to be applied to each element of X
}
\description{
lapply returns a list of the same length as X, each element of which is the result of applying FUN to the corresponding element of X.
}
\examples{
 
sp=spc.example_spectra()
BL=spc.makeSpcList(sp,"CAST")
#Counts rows (returns a list object)
spc.lapply(BL,function(x) {nrow(x)})
#Perform arithmetic operations on all Spectra elements. Returns a SpcList object.
spc.lapply(BL,function(x) {x^2+1})
}
