% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{spc.lines}
\alias{spc.lines}
\alias{spc.lines,Spectra-method}
\title{Add spectra to an existing plot}
\usage{
spc.lines(x,...)

\S4method{spc.lines}{Spectra}(x, ...)
}
\arguments{
\item{x}{An object of class \code{Spectra}}

\item{...}{Additional input arguments to be passed to lines()}
}
\description{
Adds spectra to an existing plot created by spc.plot() using lines()
}
\examples{
sp = spc.example_spectra()
spc.plot(sp[2,])
spc.lines(sp[3,],col="red")

}
\seealso{
\code{\link{spc.plot}}
}
