% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.makeSpcList}
\alias{spc.makeSpcList}
\title{Conversion from \code{Spectra} to \code{Spclist}}
\usage{
spc.makeSpcList(myobj, name)
}
\arguments{
\item{myobj}{a \code{Spectra} object}

\item{name}{name of station of a \code{Spectra} object}
}
\description{
Conversion from \code{Spectra} to \code{Spclist} using a data field
}
\examples{
sp <- spc.example_spectra()
BL = spc.makeSpcList(sp,"CAST")
show(BL)

}
