% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\name{spc.plot.depth.overlay}
\alias{spc.plot.depth.overlay}
\title{Plotting \code{SpcList} object}
\usage{
spc.plot.depth.overlay(object, X, lab_cex, ...)
}
\arguments{
\item{object}{a \code{SpcList} data}

\item{X}{column name or index}

\item{...}{any further arguments of plot}

\item{lab_cex}{vector of character expansion sizes, used cyclically}
}
\description{
Generating plot of the contents of a \code{SpcList} object overlay with respect to depth
}
\examples{
sp <- spc.example_spectra()
BL = spc.makeSpcList(sp,"CAST")
spc.plot.depth.overlay(BL, "anap_555")


}
