% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{spc.plot.depth}
\alias{spc.plot.depth}
\alias{spc.plot.depth,Spectra-method}
\title{Plotting \code{Spectra} object}
\usage{
spc.plot.depth(object,...)

\S4method{spc.plot.depth}{Spectra}(object, X, maxSp = 10, lab_cex, title,
  add = FALSE, xlab = NULL, ylab = NULL, ylim = NULL, xlim = NULL,
  lwd = 2, ...)
}
\arguments{
\item{...}{any further arguments of plot}

\item{x}{a \code{Spectra} data}
}
\description{
Generating plot of the contents of a \code{Spectra} object with respect to depth
}
\examples{
x <- spc.example_spectra()
spc.plot.depth(x)

}
\seealso{
\code{\link{spc.plot}}
}

