% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.plot.depth}
\alias{spc.plot.depth}
\title{Plotting \code{Spectra} object}
\usage{
spc.plot.depth(x,...)
}
\arguments{
\item{...}{any further arguments of plot}

\item{x}{a \code{Spectra} data}
}
\description{
Generating plot of the contents of a \code{Spectra} object with respect to depth
}
\examples{
x <- spc.example_spectra()
spc.plot.depth(x)

}
\seealso{
\code{\link{spc.plot}}
}
