% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\name{spc.plot.overlay}
\alias{spc.plot.overlay}
\title{Plotting \code{SpcList} object overlay}
\usage{
spc.plot(x,lab_cex=1,leg_idx=T, type="l", lty=1,lwd=1, col, ...)
}
\arguments{
\item{x}{a \code{SpcList} data}
}
\description{
Generating plot of the contents of a \code{SpcList} object overlay
}
\examples{
x <- spc.example_spectra()
BL = spc.makeSpcList(sp,"CAST")
spc.plot.overlay(BL)


}

