% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.plot.time.plotly}
\alias{spc.plot.time.plotly}
\title{Plot a Spectra object data with respect to time}
\usage{
spc.plot.time.plotly(sp, column, plot.max = 10, showlegend = FALSE,
  hoverinfo = "name", title = sp@LongName)
}
\arguments{
\item{sp}{A \code{Spectra} object}

\item{column}{number or name, default value is 10.}
}
\description{
Plot a \code{Spectra} object with respect to time
}
\examples{
spc.plot.time.plotly(sp)
spc.plot.time.plotly(sp, plot.max = 3)
spc.plot.time.plotly(sp, c("anap_450","anap_550","anap_650"))
}

