% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{spc.plot.time}
\alias{spc.plot.time}
\alias{spc.plot.time,Spectra-method}
\title{Plotting \code{Spectra} object}
\usage{
sspc.plot.time(object,Y,maxSp,xdata,lab_cex,lwd, ...)

\S4method{spc.plot.time}{Spectra}(object, Y, maxSp = 50, xdata = "time",
  lab_cex, lwd = 2, ...)
}
\arguments{
\item{object}{a \code{Spectra} data}

\item{...}{any further arguments of plot}

\item{Y}{a \code{Spectra} rows}

\item{maxSp}{maximum number of \code{Spectra} to plot}

\item{xdata}{type of data}

\item{lab_cex}{vector of character expansion sizes, used cyclically.}

\item{lwd}{vector of line widths}
}
\description{
Generating plot of the contents of a \code{Spectra} object with respect to time
}
\examples{
x <- spc.example_spectra()
spc.plot.time(x)

}
\seealso{
\code{\link{spc.plot.depth}}
}
