% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{spc.rbind,STIDF-method}
\alias{spc.rbind,STIDF-method}
\title{Combine \code{Spectra} objects by Rows}
\usage{
spc.cbind(...)
spc.rbind(...)
}
\arguments{
\item{...}{\code{Spectra} object}
}
\description{
Take a \code{Spectra} objects and combine by rows
}
\examples{
x <- spc.example_spectra()
nrow(x)  #[1] 26
x2 <- spc.rbind(x,x)
nrow(x2)  #[1] 52


}
