% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.rbind}
\alias{spc.rbind}
\title{Combine spectra Objects by Rows}
\usage{
spc.cbind(...)
spc.rbind(...)
}
\arguments{
\item{...}{spectra object}
}
\description{
Take a spectral objects and combine by rows
}
\examples{
x <- spc.example_spectra()
nrow(x)  #[1] 26
x2 <- spc.rbind(x,x)
nrow(x2)  #[1] 52



}

