% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.Read_NOMAD_v2}
\alias{spc.Read_NOMAD_v2}
\title{Read the NOMAD v2 bio-optical database}
\usage{
spc.Read_NOMAD_v2(fnm, skip.all.na.rows = TRUE)
}
\arguments{
\item{infile}{\code{character} containing the name of the input file.}
}
\value{
Returns an object of class \code{data.frame}.
}
\description{
Imports the NOMAD v2 database of the SeaBASS project. More information 
about this dataset can be found at \url{http://seabass.gsfc.nasa.gov/wiki/article.cgi?article=NOMAD}
}
\examples{
ap = spc.Read_NOMAD_v2(fnm)
class(ap)
spc.plot.plotly(ap[[4]], plot.max=15)
}
