% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\docType{methods}
\name{spc.setheader<-,list-method}
\alias{spc.setheader<-,list-method}
\title{Set a field of the @header slot of a \code{SpcList} object}
\usage{
\S4method{spc.setheader}{list}(object) <- value
}
\arguments{
\item{object}{A \code{SpcList} object.}

\item{value}{Object of class SpcList.}

\item{...}{arguments to be passed to or from other methods.}
}
\description{
Function sets or changes the value of a field in the header slot of \code{SpcList} object.
}
\examples{
sp=spc.example_spectra()
BL=spc.makeSpcList(sp,"CAST")
a=new("SpcHeader") # create new SpcHeader class
a$Longitude=123 
spc.setheader(BL[[1]]) <- a
h=spc.getheader(BL[[1]])
h

}
