% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\docType{methods}
\name{spc.setheader<-,list-method}
\alias{spc.setheader<-,list-method}
\title{Set a field of the @header slot of a \code{spclist} object}
\usage{
spc.setheader(x,name,...)<-value
}
\arguments{
\item{value}{Object of class SpcList}

\item{...}{arguments to be passed to or from other methods}

\item{x}{A \code{SpcList} object}

\item{name}{of the header field to be setted}
}
\description{
Function sets or changes the value of a field in the header slot of \code{spclist} object
}
\examples{
sp=spc.example_spectra()
BL=spc.makeSpcList(sp,"CAST")
a=new("SpcHeader") # create new SpcHeader class
a$Longitude=123 
spc.setheader(BL[[1]],"Station") <- a
h=spc.getheader(BL[[1]])
h

}
