% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.setheader<-}
\alias{spc.setheader<-}
\title{Set a field of the @header slot of a spectra object}
\usage{
spc.setheader(x,name)<-value
}
\arguments{
\item{x}{spectra object}

\item{name}{of the header field to be setted}
}
\description{
Function sets or changes the value of a field in the header slot of spectra object
}

