% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{spc.setheader<-}
\alias{spc.setheader<-}
\alias{spc.setheader<-,Spectra-method}
\title{Set a field of the @header slot of a \code{Spectra} object}
\usage{
spc.setheader(object) <- value

\S4method{spc.setheader}{Spectra}(object) <- value
}
\arguments{
\item{object}{A \code{Spectra} object}

\item{value}{Object of class SpcHeader}

\item{name}{of the header field to be setted}
}
\description{
Function sets or changes the value of a field in the header slot of \code{Spectra} object
}
\examples{
sp=spc.example_spectra()
a=new("SpcHeader") # create new SpcHeader class
a$Longitude=123 
spc.setheader(sp,"Station") <- a
sp@header

}
\seealso{
\code{\link{spc.getheader}}
}
