% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.setinvalid.idx<-}
\alias{spc.setinvalid.idx<-}
\title{Set rows of \code{Spectra} as invalid}
\usage{
spc.setinvalid.idx(object,value)
}
\arguments{
\item{value}{Logical vector}

\item{x}{A \code{Spectra} object}
}
\description{
Stores the row indexes as invalid
}
\examples{
 
sp = spc.example_spectra()
spc.getinvalid.idx(sp) #No invalid rows
vld = rep(TRUE,26)
vld[1:5]<-FALSE
spc.setinvalid.idx(sp)<-vld #Mark the first 5 rows as invalid
spc.getinvalid.idx(sp)
}
