% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.setinvalid.idx<-}
\alias{spc.setinvalid.idx<-}
\title{Set the selected spectra as invalid}
\usage{
spc.setinvalid.idx(x)
}
\arguments{
\item{x}{spectra object}
}
\description{
Function  stores the indexes as invaild variable
}

