% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.setselected.idx<-}
\alias{spc.setselected.idx<-}
\title{Set index to a spectra object}
\usage{
spc.setselected.idx(x,value)
}
\arguments{
\item{value}{index for spectra object}

\item{x}{spectra object}
}
\description{
Function set or change selection row index of a spectra object
}
\examples{
x <- spc.example_spectra()
idx=rep(FALSE,nrow(x)); 
idx[1:5]=TRUE
spc.setselected.idx(x)<-idx 
spc.plot(x)



}

