% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.setwavelengths<-}
\alias{spc.setwavelengths<-}
\title{Setting wavelenghts in a spectral object}
\usage{
spc.setwavelengths(...)
}
\arguments{
\item{...}{spectra object}
}
\description{
Function  to change or set wavelenghts insade of  a spectral object
}
\examples{
 x <- spc.example_spectra()
 spc.setwavelengths(x) <- 300:800
 show(x)
}

