% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.timeMatch}
\alias{spc.timeMatch}
\title{Match two time sequences}
\usage{
spc.timeMatch(master,searched,returnList,method,limits,report)
}
\arguments{
\item{master}{ordered sequence of variable of class \code{Spectra}}

\item{searched}{A variable of class \code{Spectra}which is searched}

\item{returnList}{Boolean; should a list be returned with all matches (TRUE), or a vector with single matches (FALSE)?}

\item{method}{Uses the simple  technique, same as of spacetime::timeMatch(),}

\item{limits}{the interval limits}

\item{report}{return character string which has information about searching results, default is False}
}
\description{
Match two time sequences for a \code{Spectra} object, where each can be intervals or instances.
}
\details{
Another version of spacetime::timeMatch(),method="over" uses the simple over technique. Same as of spacetime::timeMatch(),
method="nearest" finds the nearest measurement. Matches only one data for all elements of master
method="within" finds the measurements that are within the interval limits=c(upper,lower) (in seconds)
}
\examples{


}
