% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\docType{methods}
\name{spc.updateheader<-,list-method}
\alias{spc.updateheader<-,list-method}
\title{Update a field of the header slot of a Spclist object}
\usage{
spc.updateheader(onject,name)<-value
}
\arguments{
\item{object}{A \code{Spectra} objec}

\item{name}{of the header field to be updated}
}
\description{
Updates or changes the value of a field in the header slot of \code{Spclist} object
}
\examples{
sp=spc.example_spectra()
BL = spc.makeSpcList(sp,"STATION")
spc.getheader(BL)
spc.updateheader(BL,"Station")<-11
spc.getheader(BL)
}

