% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.updateheader<-}
\alias{spc.updateheader<-}
\title{Update the first or last part of column of named of a spectra object}
\usage{
spc.updateheader(x,name,...)
}
\arguments{
\item{...}{new elements of spectra object}

\item{x}{spectra object}

\item{name}{name of spactra object}
}
\description{
Function updates or changes the first or last part of column acording to mentioned name in  a spectra object
}

