% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.updateheader<-}
\alias{spc.updateheader<-}
\title{Update a field of the @header slot of a spectra object}
\usage{
spc.updateheader(x,name)<-value
}
\arguments{
\item{x}{spectra objec}

\item{name}{of the header field to be updated}
}
\description{
updates or changes the value of a field in the header slot of spectra object
}
\examples{
sp=spc.example_spectra()
sp@header
spc.updateheader(sp,"Station")<-11
sp@header
}

