% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\name{spc.updateheader<-}
\alias{spc.updateheader<-}
\title{Update a field of the @header slot of a \code{Spectra} object}
\usage{
spc.updateheader(x,name,...)<-value
}
\arguments{
\item{...}{arguments to be passed to or from other methods}

\item{x}{A \code{Spectra} objec}

\item{name}{of the header field to be updated}
}
\description{
Updates or changes the value of a field in the header slot of \code{Spectra} object
}
\examples{
sp=spc.example_spectra()
sp@header
spc.updateheader(sp,"Station")<-11
sp@header
}
