% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcHeader-Class.R
\docType{class}
\name{SpcHeader-class}
\alias{SpcHeader-class}
\title{\code{SpcHeader} class.}
\description{
Definition for \code{SpcHeader}. This class is required
for the @header slot of \code{Spectra} object.
}
\examples{
new("SpcHeader")
}
