% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcHeader-Class.R
\docType{class}
\name{SpcHeader-class}
\alias{SpcHeader-class}
\title{\code{SpcHeader} class for header object storing metadata.}
\description{
Definition for \code{SpcHeader}. This class is required
for the @header slot of \code{Spectra} object. This class directly inherits
R lists, so there is no additional slots.
}
\examples{
new("SpcHeader")
}
