% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\name{SpcList}
\alias{SpcList}
\title{\code{SpcList} class.}
\usage{
SpcList(x)
}
\arguments{
\item{x}{a list object}
}
\description{
Definition for \code{SpcList}. This class provides
a collection of multiple \code{Spectra} objects inside a list.
}
\examples{
sp=spc.example_spectra()
as(list(sp,sp^2), "SpcList")
}
