% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpcList-Methods.R
\name{SpcList}
\alias{SpcList}
\title{Constructor function for the \code{SpcList} class.}
\usage{
SpcList(x)
}
\arguments{
\item{x}{a list object}
}
\description{
With this function, it is easy to create a \code{SpcList} object
given an list containing multipls \code{Spectra} objects.
}
\examples{
sp=spc.example_spectra()
#Create an SpcList object using two Spectra objects
as(list(sp,sp^2), "SpcList")
#the above is the same as 
SpcList(list(sp,sp^2))

}
