% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Class.R
\docType{class}
\name{Spectra-class}
\alias{Spectra-class}
\title{Create a spectra class definition}
\usage{
setClass(class, contains,pepresentation,prototype)
}
\arguments{
\item{contains}{STIDF}

\item{class}{character string name for the class.}

\item{pepresentation}{avoided}

\item{prototype}{n object providing the default data for the slots in this class}
}
\description{
Create a spectra class definition, specifying the representation (the slots) and/or the classes contained in this one (the superclasses),
plus other optional details. As a side effect, the class definition is stored in the specified environment
}
