% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-Methods.R
\docType{methods}
\name{[,Spectra-method}
\alias{[,Spectra-method}
\title{Extract or replace parts of a \code{Spectra} object}
\usage{
x[i]
x[i,j]
}
\arguments{
\item{x}{A \code{Spectra} object from which to extract element(s) or in which to replace element(s).}

\item{i, j}{indices specifying elements to extract or replace. Indices are numeric or character vectors}
}
\description{
Operators acting on \code{Spectra} object and \code{Spectra} lists to extract or replace parts.
}
\details{
These operators are generic. You can write methods to handle indexing of specific classes of objects
}
\examples{
sp=spc.example_spectra()
sp #501 spectral channels in columns and 26 observations in rows 
sp[1] #501 spectral channels in columns and 1 observations in rows 

}
